Test 코드 수정

+++ src/main/java/asc/portfolio/ascSb/web/dto/seatReservationInfo/SeatReservationInfoSelectResponseDto.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 public class SeatReservationInfoSelectResponseDto {
     private Integer seatNumber;
     private Integer startTime;
-    private Integer timeInUse;
+    private Long timeInUse;
 
     private LocalDateTime createDate;
